Art Build #1
Added floating arena model with addition assets to build the arena. Also added Tal proxy asset, which can be used to size up the stage compared to the size of the character. It can also be used to add the basic controls to the characters. Added some standard asset folder, which has some of the pro scripts of unity that have really good image effect filters for camera, and other useful things.